# xxhash: use ENV.O3 for performance