big performance on mac m1
run `xxhsum -b` and compare

with -Os (homebrew default)
```
 1#XXH32                         :     102400 ->    71162 it/s ( 6949.4 MB/s)
 3#XXH64                         :     102400 ->   142409 it/s (13907.1 MB/s)
 5#XXH3_64b                      :     102400 ->   113125 it/s (11047.4 MB/s)
11#XXH128                        :     102400 ->   112679 it/s (11003.8 MB/s)
```
with -O3
```
 1#XXH32                         :     102400 ->    71310 it/s ( 6963.8 MB/s)
 3#XXH64                         :     102400 ->   142089 it/s (13875.9 MB/s)
 5#XXH3_64b                      :     102400 ->   365165 it/s (35660.6 MB/s)
11#XXH128                        :     102400 ->   364800 it/s (35625.0 MB/s)
```
that's 3x faster

and -O3 is default in https://github.com/Cyan4973/xxHash/blob/dbea33e47e7c0fe0b7c8592cd931c7430c1f130d/Makefile#L40

<!-- Use [x] to mark item done, or just click the checkboxes with device pointer -->

- [x] Have you followed the [guidelines for contributing](https://github.com/Homebrew/homebrew-core/blob/HEAD/CONTRIBUTING.md)?
- [x] Have you ensured that your commits follow the [commit style guide](https://docs.brew.sh/Formula-Cookbook#commit)?
- [x] Have you checked that there aren't other open [pull requests](https://github.com/Homebrew/homebrew-core/pulls) for the same formula update/change?
- [x] Have you built your formula locally with `HOMEBREW_NO_INSTALL_FROM_API=1 brew install --build-from-source <formula>`, where `<formula>` is the name of the formula you're submitting?
- [x] Is your test running fine `brew test <formula>`, where `<formula>` is the name of the formula you're submitting?
- [x] Does your build pass `brew audit --strict <formula>` (after doing `HOMEBREW_NO_INSTALL_FROM_API=1 brew install --build-from-source <formula>`)? If this is a new formula, does it pass `brew audit --new <formula>`?

-----