@carlocab 2024-08-21
Thanks for investigating! I agree with your assessment that the dependent failure is not related.